🎨Combine GraphQL queries (#66)
This PR combines GraphQL queries to cut down on network requests. It also destructures siteConfig when it's required to explicitly show which options are being used and introduces async/await to cut down on indentation. I grouped the templates & data extraction together as it felt cleaner, but I can revert it if needed.